# server-openapi.yml
openapi: 3.0.3
info:
  title: Ticket Backend API (Minimal)
  version: 1.0.0
  description: Minimal API for search -> order -> pay -> ticket.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /routes:
    get:
      summary: Search routes
      operationId: searchRoutes
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string }
          example: SPB
        - name: to
          in: query
          required: true
          schema: { type: string }
          example: MSK
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-03-01"
        - name: from_terminal
          in: query
          required: false
          schema: { type: string }
          example: "Moskovsky"
        - name: to_terminal
          in: query
          required: false
          schema: { type: string }
          example: "Leningradsky"

        - name: min_price
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          example: 1000
        - name: max_price
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          example: 3000

        - name: departure_after
          in: query
          required: false
          schema:
            type: string
            pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "08:00"

        - name: departure_before
          in: query
          required: false
          schema:
            type: string
            pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "18:30"

        - name: only_with_seats
          in: query
          required: false
          schema:
            type: boolean
          example: true

        - name: sort
          in: query
          required: false
          description: Sorting in format field:direction
          schema:
            type: string
            enum:
              - "price:asc"
              - "price:desc"
              - "departureTime:asc"
              - "departureTime:desc"
              - "freeSeats:asc"
              - "freeSeats:desc"
          example: "price:asc"
      responses:
        "200":
          description: Routes found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutesResponse"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /orders:
    post:
      summary: Create order (reserve seat for passenger)
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreatedResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "409":
          description: Seat already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"

  /orders/{orderId}/pay:
    post:
      summary: Pay for order
      description: |
        Payment system stub
      operationId: payOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
          example: o123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayOrderRequest"
      responses:
        "200":
          description: Payment successful; ticket issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayOrderSuccessResponse"
        "202":
          description: 3DS confirmation required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayOrderPending3dsResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "402":
          description: Payment declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentDeclinedError"
        "409":
          description: Order already paid / invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"

  /orders/{orderId}/pay/confirm-3ds:
    post:
      summary: Confirm 3DS and finalize payment
      operationId: confirmOrder3ds
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirm3dsRequest"
      responses:
        "200":
          description: Payment successful; ticket issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayOrderSuccessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "402":
          description: Payment declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentDeclinedError"
        "409":
          description: Order state invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"

  /tickets/{ticketId}:
    get:
      summary: Get ticket by ID
      operationId: getTicket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
          example: tk777
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"

components:
  schemas:
    RoutesResponse:
      type: object
      required: [routes]
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteOption"

    RouteOption:
      type: object
      required: [routeId, from, to, date, trainId, freeSeats, price]
      properties:
        routeId:
          type: string
          example: r1
        from:
          type: string
          example: SPB
        to:
          type: string
          example: MSK
        date:
          type: string
          format: date
          example: "2026-03-01"
        from_terminal:
          type: string
          nullable: true
          example: "Moskovsky"
        to_terminal:
          type: string
          nullable: true
          example: "Leningradsky"
        trainId:
          type: string
          example: t10
        freeSeats:
          type: integer
          minimum: 0
          example: 42
        price:
          type: integer
          minimum: 0
          description: Price in minor units (e.g., RUB)
          example: 1500

    CreateOrderRequest:
      type: object
      required: [routeId, seat, passenger]
      properties:
        routeId:
          type: string
          example: r1
        seat:
          type: string
          description: Seat identifier (minimal string form)
          example: "12A"
        passenger:
          $ref: "#/components/schemas/Passenger"

    Passenger:
      type: object
      required: [passportId, fullName]
      properties:
        passportId:
          type: string
          example: "1234 567890"
        fullName:
          type: string
          example: "Ivan Ivanov"

    OrderCreatedResponse:
      type: object
      required: [orderId, status, amount]
      properties:
        orderId:
          type: string
          example: o123
        status:
          type: string
          enum: [CREATED]
          example: CREATED
        amount:
          type: integer
          minimum: 0
          description: Total amount in minor units (e.g., RUB)
          example: 1500

    PayOrderRequest:
      type: object
      required: [card_number, expiration_date, cvv]
      properties:
        card_number:
          type: string
          description: PAN (учебная заглушка)
          example: "4111111111111111"
        expiration_date:
          type: string
          description: MM/YY
          example: "12/28"
        cvv:
          type: string
          description: CVV/CVC (учебная заглушка)
          example: "123"

    PayOrderSuccessResponse:
      type: object
      required: [status, ticketId]
      properties:
        status:
          type: string
          enum: [PAID]
          example: PAID
        ticketId:
          type: string
          example: tk777

    PayOrderPending3dsResponse:
      type: object
      required: [status, paymentId]
      properties:
        status:
          type: string
          enum: [PENDING_3DS]
          example: PENDING_3DS
        paymentId:
          type: string
          example: pay_123
        message:
          type: string
          nullable: true
          example: Confirm 3DS challenge

    Confirm3dsRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: "123456"

    Ticket:
      type: object
      required: [ticketId, from, to, date, trainId, seat, passenger]
      properties:
        ticketId:
          type: string
          example: tk777
        from:
          type: string
          example: SPB
        to:
          type: string
          example: MSK
        date:
          type: string
          format: date
          example: "2026-03-01"
        trainId:
          type: string
          example: t10
        seat:
          type: string
          example: "12A"
        passenger:
          $ref: "#/components/schemas/Passenger"

    ValidationError:
      type: object
      required: [type, message, details]
      properties:
        type:
          type: string
          enum: [VALIDATION_ERROR]
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Invalid request."
        details:
          type: array
          description: Field-level validation problems
          items:
            type: object
            required: [field, issue]
            properties:
              field:
                type: string
                example: date
              issue:
                type: string
                example: "must be in YYYY-MM-DD format"

    NotFoundError:
      type: object
      required: [type, resource, message]
      properties:
        type:
          type: string
          enum: [NOT_FOUND]
          example: NOT_FOUND
        resource:
          type: string
          description: What was not found
          enum: [ROUTE, ORDER, TICKET]
          example: ORDER
        message:
          type: string
          example: "Order not found."

    ConflictError:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
          enum: [CONFLICT]
          example: CONFLICT
        code:
          type: string
          description: Conflict reason
          enum: [SEAT_TAKEN, ORDER_STATE_INVALID]
          example: SEAT_TAKEN
        message:
          type: string
          example: "Selected seat is already booked."

    PaymentDeclinedError:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
          enum: [PAYMENT_DECLINED]
          example: PAYMENT_DECLINED
        code:
          type: string
          description: Decline reason from bank (minimal)
          enum: [DECLINED]
          example: DECLINED
        message:
          type: string
          example: "Payment was declined by the bank."